#10 Configured prod and dev environments with respective apiUrls + config angular.json file to use the correct environment when build for prod + at /app/services/api.service.ts changed the hard coded urls for the environment variables
