--- shop/ui/src/environments/environment.development.ts
@@ -1,4 +1,0 @@
-export const environment = {
-  production: false,
-  name: 'development',
-};
