+++ shop/ui/src/environments/environment.prod.ts
@@ -1,0 +1,6 @@
+export const environment = {
+  production: true,
+  name: 'production',
+  apiAddressUrl: 'http://portal-api-address/',
+  apiPlansUrl: 'http://portal-api-data-plans/',
+};
