+++ shop/ui/src/environments/environment.ts
@@ -1,4 +1,6 @@
 export const environment = {
-  production: true,
-  name: 'production',
+  production: false,
+  name: 'development',
+  apiAddressUrl: 'http://localhost:80/',
+  apiPlansUrl: 'http://localhost:80/',
 };
